pong/requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9
djangorestframework==3.15.1
django-cors-headers==4.3.1